added 1011S specific autonomous and some general changes
Due to our robot's unique hood, we can camp center and noone can score or descore it, unless they can cycle center, which is uncommon. We are doing a partnersship with 1011S for the high school tournie, and a specialised auto would give us the home row and perhaps even defeat 1010A and 1010G's autonomous score.
Also some other general improvements, like improving the thread that flips out our hood during autonomous, we made it a thread so it can run during the early stages of our autonomous, therefore we do not lose any time whatsoever.